+++ libs/ui/scripts/post-build.mjs
@@ -45,12 +45,12 @@ const DOC_PATHS = [
   'blocks/datatable',
   'blocks/login',
   'resources/auth',
-  'resources/generate-icons',
   'resources/global-errors',
   'resources/http-base',
   'resources/list-base',
   'resources/page-base',
   'resources/rules',
+  'icons/generate-icons',
 ];
 
 const outputDir = path.resolve('../../dist/browser');