+++ libs/ui/src/app/app.routes.ts
@@ -32,6 +32,11 @@ const localeChildren: Routes = [
     pathMatch: 'full',
   },
   {
+    path: 'icons',
+    redirectTo: 'icons/generate-icons',
+    pathMatch: 'full',
+  },
+  {
     path: 'getting-started',
     loadChildren: () => import('./features/getting-started/routes').then((m) => m.ROUTES),
   },
@@ -43,9 +48,19 @@ const localeChildren: Routes = [
     path: 'blocks',
     loadChildren: () => import('./features/blocks/routes').then((m) => m.ROUTES),
   },
+  // Legacy path from when generate-icons lived under Resources
+  {
+    path: 'resources/generate-icons',
+    redirectTo: 'icons/generate-icons',
+    pathMatch: 'full',
+  },
   {
     path: 'resources',
     loadChildren: () => import('./features/resources/routes').then((m) => m.ROUTES),
+  },
+  {
+    path: 'icons',
+    loadChildren: () => import('./features/icons/routes').then((m) => m.ROUTES),
   },
   {
     path: '**',